requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
coloredlogs==14.0
humanfriendly==8.2
idna==2.10
influxdb==5.3.0
msgpack==0.6.1
multidict==4.7.6
pyreadline==2.1
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.10
yarl==1.5.1
